クラスファイルRemoveSections.csを新規作成 クラスファイルWordMerge.CreateWordDocument.csを新規作成
