+++ src/MJS_fileJoin_src/MJS_fileJoin/MJS_fileJoin.csproj
@@ -84,6 +84,8 @@
     </Compile>
     <Compile Include="Program.cs" />
     <Compile Include="Properties\AssemblyInfo.cs" />
+    <Compile Include="RemoveSections.cs" />
+    <Compile Include="WordMerge.CreateWordDocument.cs" />
     <Compile Include="WordMerge.cs" />
     <Compile Include="WordMerge.UpdateHyperLink.cs" />
     <Compile Include="WordMerge.OutlineNumbering.cs" />
