+++ src/MJS_fileJoin_src/MJS_fileJoin/wordmerge.cs
@@ -25,7 +25,7 @@ namespace DocMergerComponent
 
             try
             {
-                (objApp,objDocLast) = CreateAndOpenWordDocument(strOrgDoc, objMissing);
+                (objApp,objDocLast) = CreateWordDocument(strOrgDoc, objMissing);
 
                 chapCnt = objDocLast.Sections.Count;
                 Dictionary<int, string> dic1 = new Dictionary<int, string>();
@@ -78,72 +78,28 @@ namespace DocMergerComponent
                 // ハイパーリンクの更新処理
                 UpdateHyperlinks(objDocLast, fm);
 
-                //fm.label10.Text = "ハイパーリンク更新中...";
-                //List<string> ls = new List<string>();
 
-                //foreach (Word.Bookmark wb in objDocLast.Bookmarks)
-                //    ls.Add(wb.Name);
+                if (check2)
+                {
+                    fm.label10.Text = "PDF出力中...";
 
-                //fm.progressBar1.Value = 0;
-                //fm.progressBar1.Maximum = objDocLast.Fields.Count;
-
-                //foreach (Word.Field wf in objDocLast.Fields)
-                //{
-                //    fm.progressBar1.Increment(1);
-                //    if (wf.Type == Word.WdFieldType.wdFieldHyperlink)
-                //    {
-                //        if (wf.Code.Text.Contains("\"http")) continue;
-                //        string text = "";
-                //        if (!wf.Code.Text.Contains(@"\l"))
-                //            text = Regex.Replace(wf.Code.Text, @".*?""([^""]*?)"".*?", "$1");
-                //        else
-                //        {
-                //            if (!Regex.IsMatch(wf.Code.Text, @".*?""[^""]*?"".*?""[^""]*?"".*?")) continue;
-                //            text = Regex.Replace(wf.Code.Text, @".*?""([^""]*?)"".*?""([^""]*?)"".*?", "$1#$2");
-                //        }
-
-                //        string[] subtext = text.Split('\\');
-                //        text = subtext[subtext.Count() - 1];
-                //        subtext = text.Split('/');
-                //        text = subtext[subtext.Count() - 1];
-                //        if (ls.Contains(text.Replace(".html", "").Replace("#", "♯").Trim()))
-                //        {
-                //            wf.Code.Text = @"HYPERLINK \l """ + text.Replace(".html", "").Replace("#", "♯").Trim() + @"""";
-                //            wf.Update();
-                //        }
-                //        else
-                //            wf.Unlink();
-                //    }
-                //}
-
-                //foreach (Word.Hyperlink wh in objDocLast.Hyperlinks)
-                //{
-                //    if (Regex.IsMatch(wh.Name.Trim(), @"^\w{3}\d{5}") ||
-                //        Regex.IsMatch(wh.Name.Trim(), @"^\w{3}\d{5}♯\w{3}\d{5}"))
-                //        wh.TextToDisplay = Regex.Replace(wh.TextToDisplay, @".*(\d+\.)+\d+[\s　\t]", "");
-                //}
-
-                //if (check2)
-                //{
-                //    fm.label10.Text = "PDF出力中...";
-
-                //    objDocLast.ExportAsFixedFormat(
-                //        strOutDoc.Replace(".doc", ".pdf"),
-                //        Word.WdExportFormat.wdExportFormatPDF,
-                //        false,
-                //        Word.WdExportOptimizeFor.wdExportOptimizeForPrint,
-                //        Word.WdExportRange.wdExportAllDocument,
-                //        1,
-                //        1,
-                //        Word.WdExportItem.wdExportDocumentContent,
-                //        false,
-                //        true,
-                //        Word.WdExportCreateBookmarks.wdExportCreateHeadingBookmarks,
-                //        false,
-                //        true,
-                //        false
-                //        );
-                //}
+                    objDocLast.ExportAsFixedFormat(
+                        strOutDoc.Replace(".doc", ".pdf"),
+                        Word.WdExportFormat.wdExportFormatPDF,
+                        false,
+                        Word.WdExportOptimizeFor.wdExportOptimizeForPrint,
+                        Word.WdExportRange.wdExportAllDocument,
+                        1,
+                        1,
+                        Word.WdExportItem.wdExportDocumentContent,
+                        false,
+                        true,
+                        Word.WdExportCreateBookmarks.wdExportCreateHeadingBookmarks,
+                        false,
+                        true,
+                        false
+                        );
+                }
 
                 if (check1)
                 {
@@ -196,43 +152,7 @@ namespace DocMergerComponent
             Merge(strOrgDoc, arrFiles, strOutDoc, form, check1, check2, check3, null);
         }
 
-        private void RemoveDuplicateSectionsByStyle(Word.Document objDocLast, MainForm fm, int chapCnt, ref int chapCntLast)
-        {
-            fm.label10.Text = "重複箇所削除中...";
-            fm.progressBar1.Maximum = 11;
-            fm.progressBar1.Value = 1;
 
-            string[] lsStyleName = { "MJS_見出し 1（項番なし）", "MJS_見出し 2（項番なし）", "MJS_マニュアルタイトル", "MJS_目次", "奥付タイトル", "索引見出し" };
-            foreach (string styleName in lsStyleName)
-            {
-                object styleObject = styleName;
-                for (int i = chapCnt + 1; i <= chapCntLast; i++)
-                {
-                    Word.Range wr = objDocLast.Sections[i].Range;
-                    wr.Find.ClearFormatting();
-
-                    if (objDocLast.Styles.Cast<Word.Style>().Any(s => s.NameLocal == styleName))
-                    {
-                        wr.Find.set_Style(ref styleObject);
-                    }
-                    else
-                    {
-                        // スタイルがなければスキップやログ出力
-                    }
-
-                    //wr.Find.Wrap = Word.WdFindWrap.wdFindStop;
-                    wr.Find.Execute();
-
-                    if (wr.Find.Found)
-                    {
-                        objDocLast.Sections[i].Range.Delete();
-                        i--;
-                        chapCntLast--;
-                    }
-                }
-                fm.progressBar1.Increment(1);
-            }
-        }
 
         private bool CheckIndexSection(Word.Document objDocLast, MainForm fm, int chapCntLast)
         {
@@ -267,53 +187,7 @@ namespace DocMergerComponent
             return last;
         }
 
-        private void RemoveLastSectionsByStyle(Word.Document objDocLast, MainForm fm, int chapCntLast, ref bool last)
-        {
-            fm.label10.Text = "章扉章節項番号修正中...";
-            fm.progressBar1.Maximum = objDocLast.Sections.Count;
-            fm.progressBar1.Value = 1;
-
-            string[] lastItems = { "MJS_見出し 1（項番なし）", "MJS_見出し 2（項番なし）", "MJS_マニュアルタイトル", "MJS_目次" };
-            foreach (string styleName in lastItems)
-            {
-                object styleObject = styleName;
-                int allChap = objDocLast.Sections.Count;
-                for (int i = allChap; i > chapCntLast; i--)
-                {
-                    try
-                    {
-                        Word.Range wr = objDocLast.Sections[i].Range;
-                        wr.Find.ClearFormatting();
-                        if (objDocLast.Styles.Cast<Word.Style>().Any(s => s.NameLocal == styleName))
-                        {
-                            wr.Find.set_Style(ref styleObject);
-                        }
-                        else
-                        {
-                            // スタイルがなければスキップやログ出力
-                        }
-                        //wr.Find.Wrap = Word.WdFindWrap.wdFindStop;
-                        wr.Find.Execute();
-                        if (wr.Find.Found)
-                        {
-                            if (last)
-                                last = false;
-                            else
-                            {
-                                objDocLast.Sections[i].Range.Delete();
-                                i--;
-                                chapCntLast--;
-                            }
-                        }
-                    }
-                    catch
-                    {
-                        break;
-                    }
-                }
-            }
-        }
-
+        
         private void UpdateChapterFrontNumbers(Word.Document objDocLast, MainForm fm)
         {
             string[] chapFrontItems = { "MJS_章扉-タイトル" };
@@ -356,59 +230,6 @@ namespace DocMergerComponent
                     catch
                     { }
                     fm.progressBar1.Increment(1);
-                }
-            }
-        }
-
-        private void ApplyOutlineNumbering(Word.Application objApp, Word.Document objDocLast, MainForm fm)
-        {
-            List<string> styleNames = new List<string>();
-            styleNames.Add("MJS_章扉-タイトル");
-            styleNames.Add("見出し 1,MJS_見出し 1");
-            styleNames.Add("見出し 2,MJS_見出し 2");
-            styleNames.Add("見出し 3,MJS_見出し 3");
-
-            // スタイルのアウトライン番号を設定
-            SetOutlineNumberingFormat(objApp, objDocLast, fm);
-
-            int first = 0;
-            int second = 0;
-            int third = 0;
-            int fourth = 0;
-
-            for (int i = 1; i <= objDocLast.ListParagraphs.Count; i++)
-            {
-                fm.progressBar1.Increment(1);
-                if (!Regex.IsMatch(objDocLast.ListParagraphs[i].Range.ListFormat.ListString, @"第.*?章") && !Regex.IsMatch(objDocLast.ListParagraphs[i].Range.ListFormat.ListString, @"\d\.\d")) continue;
-                if (Regex.IsMatch(objDocLast.ListParagraphs[i].Range.ListFormat.ListString, @"第.*?章"))
-                {
-                    first++;
-                    second = 0;
-                    third = 0;
-                    fourth = 0;
-                    if (objDocLast.ListParagraphs[i].Range.ListFormat.ListValue != first)
-                        objDocLast.ListParagraphs[i].Range.ListFormat.ApplyListTemplateWithLevel(objApp.ListGalleries[Word.WdListGalleryType.wdOutlineNumberGallery].ListTemplates[7], true, Word.WdListApplyTo.wdListApplyToWholeList, Word.WdDefaultListBehavior.wdWord10ListBehavior);
-                }
-                else if (objDocLast.ListParagraphs[i].Range.ListFormat.ListLevelNumber == 2)
-                {
-                    second++;
-                    third = 0;
-                    fourth = 0;
-                    if (objDocLast.ListParagraphs[i].Range.ListFormat.ListValue != second)
-                        objDocLast.ListParagraphs[i].Range.ListFormat.ApplyListTemplateWithLevel(objApp.ListGalleries[Word.WdListGalleryType.wdOutlineNumberGallery].ListTemplates[7], true, Word.WdListApplyTo.wdListApplyToWholeList, Word.WdDefaultListBehavior.wdWord10ListBehavior);
-                }
-                else if (objDocLast.ListParagraphs[i].Range.ListFormat.ListLevelNumber == 3)
-                {
-                    third++;
-                    fourth = 0;
-                    if (objDocLast.ListParagraphs[i].Range.ListFormat.ListValue != third)
-                        objDocLast.ListParagraphs[i].Range.ListFormat.ApplyListTemplateWithLevel(objApp.ListGalleries[Word.WdListGalleryType.wdOutlineNumberGallery].ListTemplates[7], true, Word.WdListApplyTo.wdListApplyToWholeList, Word.WdDefaultListBehavior.wdWord10ListBehavior);
-                }
-                else if (objDocLast.ListParagraphs[i].Range.ListFormat.ListLevelNumber == 4)
-                {
-                    fourth++;
-                    if (objDocLast.ListParagraphs[i].Range.ListFormat.ListValue != fourth)
-                        objDocLast.ListParagraphs[i].Range.ListFormat.ApplyListTemplateWithLevel(objApp.ListGalleries[Word.WdListGalleryType.wdOutlineNumberGallery].ListTemplates[7], true, Word.WdListApplyTo.wdListApplyToWholeList, Word.WdDefaultListBehavior.wdWord10ListBehavior);
                 }
             }
         }
